Update examples with new Sharding API
