+++ examples/example-core/config-utility/src/main/resources/META-INF/services/org.apache.shardingsphere.spi.algorithm.ShardingAlgorithm
@@ -14,8 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-driverClassName: com.mysql.jdbc.Driver
-jdbcUrl: jdbc:mysql://localhost:3307/databases_tables
-username: root
-password: root
+ 
+org.apache.shardingsphere.example.algorithm.StandardModuloShardingTableAlgorithm
+org.apache.shardingsphere.example.algorithm.StandardModuloShardingDatabaseAlgorithm
