+++ examples/sharding-proxy-example/sharding-proxy-hint-example/src/main/java/org/apache/shardingsphere/example/proxy/hint/ModuloHintShardingAlgorithm.java
@@ -22,6 +22,7 @@ import org.apache.shardingsphere.api.sharding.hint.HintShardingValue;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Properties;
 
 public final class ModuloHintShardingAlgorithm implements HintShardingAlgorithm<String> {
     
@@ -36,5 +37,19 @@ public final class ModuloHintShardingAlgorithm implements HintShardingAlgorithm<
             }
         }
         return result;
+    }
+    
+    @Override
+    public String getType() {
+        return "HINT_TEST";
+    }
+    
+    @Override
+    public Properties getProperties() {
+        return new Properties();
+    }
+    
+    @Override
+    public void setProperties(final Properties properties) {
     }
 }
